# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Club Event & Fee Management API
  version: 0.1.0
  description: 교내 개발 동아리 행사·회비 관리 웹앱의 첫 번째 세로 기능 API 계약
servers:
  - url: /api/v1
security:
  - cookieAuth: []
...
paths:
...
  /events/{eventId}:
    get:
      tags: [Events]
      operationId: getEvent
      summary: 행사 상세 조회
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: 행사 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/dashboard:
    get:
      tags: [Admin]
      operationId: getAdminDashboard
      summary: 운영진 대시보드 조회
      responses:
        '200':
          description: 운영 현황 요약
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDashboardResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /admin/events:
    get:
      tags: [Admin]
      operationId: getAdminEvents
      summary: 운영진 행사 전체 목록 조회
      responses:
        '200':
          description: 행사 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminEventResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Admin]
      operationId: createAdminEvent
      summary: 행사 초안 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminEventCreateRequest'
      responses:
        '201':
          description: 생성된 행사 초안
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminEventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/events/{eventId}:
    get:
      tags: [Admin]
      operationId: getAdminEvent
      summary: 운영진 행사 상세 조회
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: 행사 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminEventResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Admin]
      operationId: updateAdminEvent
      summary: 행사 초안 수정
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminEventUpdateRequest'
      responses:
        '200':
          description: 수정된 행사
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminEventResponse'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Admin]
      operationId: deleteAdminEvent
      summary: 참가 이력이 없는 행사 초안 삭제
      parameters:
        - $ref: '#/components/parameters/EventId'
        - name: version
          in: query
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '204':
          description: 삭제 완료
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/events/{eventId}/publish:
    post:
      tags: [Admin]
      operationId: publishAdminEvent
      summary: 행사 초안 공개
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminEventVersionRequest'
      responses:
        '200':
          description: 공개된 행사
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminEventResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/events/{eventId}/close:
    post:
      tags: [Admin]
      operationId: closeAdminEvent
      summary: 공개 행사 종료
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminEventTransitionRequest'
      responses:
        '200':
          description: 종료된 행사
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminEventResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/events/{eventId}/cancel:
    post:
      tags: [Admin]
      operationId: cancelAdminEvent
      summary: 행사 취소 및 납부 상태 정리
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminEventTransitionRequest'
      responses:
        '200':
          description: 행사 취소 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminEventCancelResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/events/{eventId}/participants:
    get:
      tags: [Admin]
      operationId: getAdminEventParticipants
      summary: 행사별 참가자와 참가비 납부 현황 조회
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: paymentStatus
          in: query
          schema:
            $ref: '#/components/schemas/PaymentStatus'
      responses:
        '200':
          description: 행사 참가자 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminEventParticipantPage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/dues-rounds/{duesRoundId}/payments:
    get:
      tags: [Admin]
      operationId: getAdminDuesPayments
      summary: 회비 차수별 부원 납부 현황 조회
      parameters:
        - $ref: '#/components/parameters/DuesRoundId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: paymentStatus
          in: query
          schema:
            $ref: '#/components/schemas/PaymentStatus'
      responses:
        '200':
          description: 회비 납부 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDuesPaymentPage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/dues-rounds:
    get:
      tags: [Admin]
      operationId: getAdminDuesRounds
      summary: 회비 차수 목록 조회
      responses:
        '200':
          description: 회비 차수 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminDuesRoundResponse'
    post:
      tags: [Admin]
      operationId: createAdminDuesRound
      summary: 회비 차수 초안 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminDuesRoundRequest'
      responses:
        '201':
          description: 생성된 회비 차수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDuesRoundResponse'
  /admin/dues-rounds/{duesRoundId}:
    patch:
      tags: [Admin]
      operationId: updateAdminDuesRound
      summary: 회비 차수 초안 수정
      parameters:
        - $ref: '#/components/parameters/DuesRoundId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminDuesRoundRequest'
      responses:
        '200':
          description: 수정된 회비 차수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDuesRoundResponse'
    delete:
      tags: [Admin]
      operationId: deleteAdminDuesRound
      summary: 회비 차수 초안 삭제
      parameters:
        - $ref: '#/components/parameters/DuesRoundId'
        - name: version
          in: query
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '204':
          description: 삭제 완료
  /admin/dues-rounds/{duesRoundId}/publish:
    post:
      tags: [Admin]
      operationId: publishAdminDuesRound
      summary: 회비 차수 공개 및 활성 부원 납부 항목 생성
      parameters:
        - $ref: '#/components/parameters/DuesRoundId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminEventVersionRequest'
      responses:
        '200':
          description: 공개 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDuesPublishResponse'
  /admin/payment-obligations/{paymentId}/confirm:
    post:
      tags: [Admin]
      operationId: confirmAdminPayment
      summary: 송금 신고 승인
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPaymentReviewRequest'
      responses:
        '200':
          description: 승인된 납부 상태
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPaymentReviewResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/payment-obligations/{paymentId}/reject:
    post:
      tags: [Admin]
      operationId: rejectAdminPayment
      summary: 송금 신고 반려
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPaymentReviewRequest'
      responses:
        '200':
          description: 반려된 납부 상태
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPaymentReviewResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /admin/payment-obligations/{paymentId}/refund:
    post:
      tags: [Admin]
      operationId: completeAdminRefund
      summary: 환불 완료 처리
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPaymentReviewRequest'
      responses:
        '200':
          description: 환불 완료 상태
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPaymentReviewResponse'
...
  /admin/members/{memberId}/approve:
    post:
      tags: [Admin]
      operationId: approveAdminMember
      summary: 가입 대기 회원 승인
      parameters:
        - $ref: '#/components/parameters/MemberId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminMemberActionRequest'
      responses:
        '200':
          description: 승인된 회원
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMemberResponse'
  /admin/members/{memberId}/suspend:
    post:
      tags: [Admin]
      operationId: suspendAdminMember
      summary: 활성 회원 정지
      parameters:
        - $ref: '#/components/parameters/MemberId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminMemberActionRequest'
      responses:
        '200':
          description: 정지된 회원
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMemberResponse'
  /admin/members/{memberId}/role:
    patch:
      tags: [Admin]
      operationId: changeAdminMemberRole
      summary: 활성 회원 역할 변경
      parameters:
        - $ref: '#/components/parameters/MemberId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminMemberRoleRequest'
      responses:
        '200':
          description: 역할이 변경된 회원
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMemberResponse'
  /admin/payment-settings:
    get:
      tags: [Admin]
      operationId: getAdminPaymentSettings
      summary: 송금정보 버전 목록
      responses:
        '200':
          description: 송금정보 버전 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentSettingResponse'
    post:
      tags: [Admin]
      operationId: createAdminPaymentSetting
      summary: 새 송금정보 등록 및 활성화
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentSettingRequest'
      responses:
        '201':
          description: 활성화된 송금정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentSettingResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SESSION
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    PaymentId:
      name: paymentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    DuesRoundId:
      name: duesRoundId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    MemberId:
      name: memberId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Forbidden:
      description: 접근 권한 없음
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: 리소스 없음
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: 현재 상태와 요청 충돌
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    MemberPart:
      type: string
      enum: [PLAN, DESIGN, PE_WEB, PE_MOBILE]
...
    EventStatus:
      type: string
      enum: [DRAFT, PUBLISHED, CLOSED, CANCELED]
...
    PaymentStatus:
      type: string
      enum: [NOT_REQUIRED, UNPAID, REPORTED, CONFIRMED, REJECTED, VOID, REFUND_PENDING, REFUNDED]
...
    EventListItem:
      type: object
      required: [id, title, startsAt, registrationDeadline, joinedCount, feeAmount, status]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        summary:
          type: string
        location:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        registrationDeadline:
          type: string
          format: date-time
        capacity:
          type: integer
        joinedCount:
          type: integer
        feeAmount:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/EventStatus'
        myParticipationStatus:
          $ref: '#/components/schemas/ParticipationStatus'
        myPaymentStatus:
          $ref: '#/components/schemas/PaymentStatus'
...
    ParticipationSummary:
      type: object
      required: [id, status, joinedAt, version]
      properties:
        id:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/ParticipationStatus'
        joinedAt:
          type: string
          format: date-time
        canceledAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
...
    EventDetail:
      allOf:
        - $ref: '#/components/schemas/EventListItem'
        - type: object
          required: [description, allowLateCancellation, canJoin, canCancel]
          properties:
            description:
              type: string
            allowLateCancellation:
              type: boolean
            canJoin:
              type: boolean
            canCancel:
              type: boolean
            myParticipation:
              $ref: '#/components/schemas/ParticipationSummary'
            myPayment:
              $ref: '#/components/schemas/PaymentSummary'
...
    AdminEventCreateRequest:
      type: object
      required: [title, description, startsAt, registrationDeadline, feeAmount, allowLateCancellation]
      properties:
        title:
          type: string
          maxLength: 200
        summary:
          type: string
          maxLength: 500
        description:
          type: string
        location:
          type: string
          maxLength: 200
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        registrationDeadline:
          type: string
          format: date-time
        capacity:
          type: integer
          minimum: 1
        feeAmount:
          type: integer
          format: int64
          minimum: 0
        allowLateCancellation:
          type: boolean
    AdminEventUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/AdminEventCreateRequest'
        - type: object
          required: [version]
          properties:
            version:
              type: integer
              format: int64
              minimum: 0
    AdminEventVersionRequest:
      type: object
      required: [version]
      properties:
        version:
          type: integer
          format: int64
          minimum: 0
    AdminEventTransitionRequest:
      type: object
      required: [version]
      properties:
        version:
          type: integer
          format: int64
          minimum: 0
        reason:
          type: string
          maxLength: 500
    AdminEventCancelResponse:
      type: object
      required: [eventStatus, voidedPaymentCount, refundPendingCount, version]
      properties:
        eventStatus:
          $ref: '#/components/schemas/EventStatus'
        voidedPaymentCount:
          type: integer
          format: int64
        refundPendingCount:
          type: integer
          format: int64
        version:
          type: integer
          format: int64
    AdminEventResponse:
      type: object
      required: [id, title, description, startsAt, registrationDeadline, joinedCount, feeAmount, status, allowLateCancellation, createdAt, updatedAt, version]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        summary:
          type: string
        description:
          type: string
        location:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        registrationDeadline:
          type: string
          format: date-time
        capacity:
          type: integer
        joinedCount:
          type: integer
          format: int64
        feeAmount:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/EventStatus'
        allowLateCancellation:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
    AdminDashboardResponse:
      type: object
      required: [memberCount, pendingMemberCount, upcomingEventCount, unpaidCount, reportedCount, expectedAmount, confirmedAmount, collectionRate, upcomingEvents, activeDuesRounds, recentReports]
      properties:
        memberCount:
          type: integer
        pendingMemberCount:
          type: integer
        upcomingEventCount:
          type: integer
        unpaidCount:
          type: integer
        reportedCount:
          type: integer
        expectedAmount:
          type: integer
          format: int64
        confirmedAmount:
          type: integer
          format: int64
        collectionRate:
          type: number
          format: double
          minimum: 0
          maximum: 100
        upcomingEvents:
          type: array
          items:
            $ref: '#/components/schemas/AdminEventSummary'
        activeDuesRounds:
          type: array
          items:
            $ref: '#/components/schemas/AdminDuesRoundSummary'
        recentReports:
          type: array
          items:
            $ref: '#/components/schemas/AdminPaymentRow'
...
    AdminDuesRoundStatus:
      type: string
      enum: [DRAFT, PUBLISHED, CLOSED]
    AdminDuesRoundRequest:
      type: object
      required: [title, amount, dueAt, version]
      properties:
        title:
          type: string
          maxLength: 200
        amount:
          type: integer
          format: int64
          minimum: 0
        dueAt:
          type: string
          format: date-time
        bankName:
          type: string
          maxLength: 100
        accountNumber:
          type: string
          maxLength: 100
        accountHolder:
          type: string
          maxLength: 100
        kakaoPayReceiveUrl:
          type: string
          maxLength: 500
        version:
          type: integer
          format: int64
          minimum: 0
    AdminDuesRoundResponse:
      type: object
      required: [id, title, amount, dueAt, status, targetCount, createdAt, updatedAt, version]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        amount:
          type: integer
          format: int64
        dueAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/AdminDuesRoundStatus'
        bankName:
          type: string
        accountNumber:
          type: string
        accountHolder:
          type: string
        kakaoPayReceiveUrl:
          type: string
        targetCount:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
    AdminDuesPublishResponse:
      type: object
      required: [duesRound, createdPaymentCount]
      properties:
        duesRound:
          $ref: '#/components/schemas/AdminDuesRoundResponse'
        createdPaymentCount:
          type: integer
          format: int64
...
    AdminEventParticipantPage:
      type: object
      required: [event, items, page, size, totalElements, totalPages]
      properties:
        event:
          $ref: '#/components/schemas/AdminEventSummary'
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminEventParticipant'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    AdminDuesPaymentPage:
      type: object
      required: [duesRound, items, page, size, totalElements, totalPages]
      properties:
        duesRound:
          $ref: '#/components/schemas/AdminDuesRoundSummary'
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminPaymentRow'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    AdminPaymentReviewRequest:
      type: object
      required: [version]
      properties:
        version:
          type: integer
          format: int64
        note:
          type: string
          maxLength: 500
    AdminPaymentReviewResponse:
      type: object
      required: [paymentId, status, version, reviewedAt]
      properties:
        paymentId:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/PaymentStatus'
        version:
          type: integer
          format: int64
        reviewedAt:
          type: string
          format: date-time
    AdminMemberResponse:
      type: object
      required: [id, kakaoProfileName, displayNickname, role, status, onboardingCompleted, createdAt, updatedAt, version]
      properties:
        id:
          type: integer
          format: int64
        kakaoProfileName:
          type: string
        name:
          type: string
        part:
          $ref: '#/components/schemas/MemberPart'
        displayNickname:
          type: string
        role:
          $ref: '#/components/schemas/MemberRole'
        status:
          $ref: '#/components/schemas/MemberStatus'
        onboardingCompleted:
          type: boolean
        approvedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
    AdminMemberActionRequest:
      type: object
      required: [version]
      properties:
        version:
          type: integer
          format: int64
          minimum: 0
    AdminMemberRoleRequest:
      type: object
      required: [role, version]
      properties:
        role:
          $ref: '#/components/schemas/MemberRole'
        version:
          type: integer
          format: int64
          minimum: 0
    PaymentSettingRequest:
      type: object
      required: [bankName, accountNumber, accountHolder]
      properties:
        bankName:
          type: string
          maxLength: 100
        accountNumber:
          type: string
          maxLength: 100
        accountHolder:
          type: string
          maxLength: 100
        kakaoPayReceiveUrl:
          type: string
          maxLength: 500
    PaymentSettingResponse:
      type: object
      required: [id, bankName, accountNumber, accountHolder, active, createdBy, createdAt]
      properties:
        id:
          type: integer
          format: int64
        bankName:
          type: string
        accountNumber:
          type: string
        accountHolder:
          type: string
        kakaoPayReceiveUrl:
          type: string
        active:
          type: boolean
        createdBy:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
...